More changes
Added [In, Out] attribute for Interops functions
Implemented AllocateDescriptorSets

+++ Magnesium.Vulkan/Structs/VkDescriptorSetAllocateInfo.cs
@@ -11,6 +11,6 @@ namespace Magnesium.Vulkan
 		public IntPtr pNext { get; set; }
 		public UInt64 descriptorPool { get; set; }
 		public UInt32 descriptorSetCount { get; set; }
-		public UInt64 pSetLayouts { get; set; }
-	}
+		public IntPtr pSetLayouts { get; set; } // UInt64
+}
 }
